requirements.txt
Flask==3.0.3
google-generativeai
python-dotenv==1.0.1
gunicorn

